remove jetty6 support, migrate needed jetty6 bits to GWTHostedMode, update jetty support for jetty 7.2+, 8 and 9
