--- environments/servlet/core/src/main/java/org/jboss/weld/environment/jetty/Jetty6Container.java
@@ -1,39 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2010, Red Hat, Inc., and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.weld.environment.jetty;
-
-import org.jboss.weld.environment.Container;
-
-/**
- * Jetty6 container.
- *
- * @author <a href="mailto:ales.justin@jboss.org">Ales Justin</a>
- */
-public class Jetty6Container extends AbstractJettyPre72Container {
-    public static Container INSTANCE = new Jetty6Container();
-
-    private static final String JETTY_REQUIRED_CLASS_NAME = "org.mortbay.jetty.servlet.ServletHandler";
-
-    protected String classToCheck() {
-        return JETTY_REQUIRED_CLASS_NAME;
-    }
-
-    protected Class<?> getWeldServletHandlerClass() {
-        return MortbayWeldServletHandler.class;
-    }
-}
