--- environments/servlet/tests/jetty6/src/test/debug-resources/META-INF/jboss-test-harness.properties
@@ -1,7 +1,0 @@
-# Configuration for running incontainer tests from your IDE
-# Alter the path weld accordingly (relative from the tck/impl dir)
-org.jboss.testharness.standalone=false
-org.jboss.testharness.container.extraConfigurationDir=../
-org.jboss.testharness.container.forceRestart=false
-org.jboss.testharness.libraryDirectory=target/dependency/lib
-org.jboss.testharness.runIntegrationTests=true
