--- environments/servlet/tests/jetty6/src/test/java/org/jboss/weld/environment/servlet/test/config/ConfigTest.java
@@ -1,38 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat, Inc., and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.weld.environment.servlet.test.config;
-
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.jboss.weld.environment.servlet.test.util.JettyDeployments;
-import org.junit.runner.RunWith;
-
-/**
- * @author <a href="mailto:ales.justin@jboss.org">Ales Justin</a>
- */
-@RunWith(Arquillian.class)
-public class ConfigTest extends ConfigTestBase {
-
-    @Deployment
-    public static WebArchive getDeployment() {
-        WebArchive war = baseDOSDeployment();
-        war.addAsWebInfResource(JettyDeployments.JETTY_ENV, "jetty-env.xml");
-        return war;
-    }
-}
