--- environments/servlet/tests/jetty6/src/test/java/org/jboss/weld/environment/servlet/test/injection/LookupTest.java
@@ -1,19 +1,0 @@
-package org.jboss.weld.environment.servlet.test.injection;
-
-
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.runner.RunWith;
-
-import static org.jboss.weld.environment.servlet.test.util.JettyDeployments.JETTY_ENV;
-
-@RunWith(Arquillian.class)
-public class LookupTest extends LookupTestBase {
-
-    @Deployment
-    public static WebArchive deployment() {
-        return LookupTestBase.deployment().addAsWebInfResource(JETTY_ENV, "jetty-env.xml");
-    }
-
-}
