--- environments/servlet/tests/jetty6/src/test/resources/arquillian.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<arquillian xmlns="http://www.jboss.org/arquillian-1.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.jboss.org/arquillian-1.0 http://www.jboss.org/schema/arquillian/arquillian-1.0.xsd">
-
-	<engine>
-		<property name="deploymentExportPath">target/</property>
-	</engine>
-
-</arquillian>
